Refactor ThemedApp
